ui/requirements.txt
allure-pytest==2.5.2
PyHamcrest==1.9.0
pytest==3.8.1
pytest-xdist==1.23.1
PyYAML==3.13
requests==2.19.1
selenium==3.14.1
